+++ assets/js/lib/ReviewAPI.js
@@ -1,0 +1,9 @@
+const fetchReviews = async () => {
+  const { data: reviewData } = await fetch("/web_api/reviews").then((res) =>
+    res.json(),
+  );
+
+  return reviewData;
+};
+
+export { fetchReviews };
